Refactor Piece trait to piece struct
Piece is now build with a piece_type field that can be queried by owners
of pieces. This field also determines the behavior of get_moves. It's a
deviation from object-oriented styles, but it was important to be able
to query a piece to know type info about it, e.g. if it was the King.